admin functionalities added * import movies , users, ratings in database * display top trending movies or popular movies of particular genre

+++ src/main/program_admin.py
@@ -1,9 +1,9 @@
 import mongoengine
 import pandas as pd
-from src.lib.driver import *
-from src.lib.switchlang import switch
 from src.account import account as acc
 from src.data import service
+from src.lib.driver import *
+from src.lib.switchlang import switch
 
 
 def run():
@@ -15,12 +15,19 @@ def run():
 
         with switch(action) as s:
             s.case('a', add_movie)
-            s.case('al', add_movie_from_file)
+            s.case('lm', load_movies)
+            s.case('lu', load_users)
+            s.case('lr', load_ratings)
+
             s.case('s', service.display_all_movies)
+            s.case('info', service.user_info)
+            s.case('users', service.user_count)
+            s.case('movies', service.movie_count)
+            s.case('ratings', service.rating_count)
 
             s.case('r', db_reset)
-            s.case('c', acc.create_account)
-            s.case('l', acc.log_into_account)
+            s.case('c', acc.create_admin_account)
+            s.case('l', acc.log_into_admin)
             s.case('m', lambda: 'change_mode')
             s.case('', lambda: None)
             s.case(['x', 'bye', 'exit', 'exit()'], exit_app)
@@ -39,7 +46,9 @@ def run():
 def show_commands():
     print('What action would you like to take:')
     print('[A]dd a movie')
-    print('[AL]dd a movie from a csv file')
+    print('[lm]Add movies from a csv file')
+    print('[lu]Add users from a csv file')
+    print('[lr]Add ratings from a csv file')
     print('[S]how all movies')
     print('[L]og into account')
     print('[C]reate an account')
@@ -52,30 +61,35 @@ def show_commands():
 def add_movie():
     print(' ************   Add a movie   **************** ')
     if not acc.active_account:
-        error_msg("You must log in first to review a movie")
+        error_msg("You must log in first to add a movie")
         return
-
-    movie_id = input("What will be the  movie ID")
+    if not acc.active_account.admin:
+        error_msg('Only admin can use this feature ')
+        return
     name = input("What is the name of movie")
-    movie = service.add_movie(movie_id, name)
+    movie = service.add_movie(name)
     if movie.id:
         success_msg("Successfuly added movie {}".format(movie.name))
 
 
-def add_movie_from_file():
-    print(' ************   Add a movie   **************** ')
+def load_movies():
+    print(' ************   Add Movies  **************** ')
     if not acc.active_account:
-        error_msg("You must log in first to review a movie")
+        error_msg("You must log in first to add movies")
+        return
+    if not acc.active_account.admin:
+        error_msg('Only admin can use this feature ')
         return
 
     while True:
         try:
             file_path = input("Enter full path of movie file")
             data = pd.read_csv(file_path)
-            data = data.rename(columns={"movieId": "movie_id", "title": "name"})
+            data = data.rename(columns={"title": "name", "movieId": "id"})
             data_list = data.to_dict('records')
-            for dict in data_list:
-                dict['genres'] = dict['genres'].split('|')
+            for x in data_list:
+                x['genres'] = x['genres'].split('|')
+            print('************loading data***************')
 
             service.add_movie_from_list(data_list)
             break
@@ -83,6 +97,53 @@ def add_movie_from_file():
             print(" {}  OR press M to go back to menu : ".format(e))
             if file_path == 'b' or file_path == 'B':
                 break
+
+
+def load_users():
+    print(' ************   Load Users  **************** ')
+    if not acc.active_account:
+        error_msg("You must log in first")
+        return
+    if not acc.active_account.admin:
+        error_msg('Only admin can use this feature ')
+        return
+    while True:
+        try:
+            file_path = input("Enter full path of users file")
+            data = pd.read_csv(file_path)
+            data_list = data.to_dict('records')
+            print('************loading data***************')
+
+            acc.create_accounts_from_list(data_list)
+            break
+        except Exception as e:
+            print(" {}  OR press M to go back to menu : ".format(e))
+            if file_path == 'b' or file_path == 'B':
+                break
+
+
+def load_ratings():
+    print(' ************   Loading Ratings  **************** ')
+    if not acc.active_account:
+        error_msg("You must log in first")
+        return
+    if not acc.active_account.admin:
+        error_msg('Only admin can use this feature ')
+        return
+    while True:
+        try:
+            file_path = input("Enter full path of ratings file")
+            data = pd.read_csv(file_path)
+            data_list = data.to_dict('records')
+            print('************loading data***************')
+
+            service.add_ratings_from_list(data_list)
+            break
+        except Exception as e:
+            print(" {}  OR press M to go back to menu : ".format(e))
+            if file_path == 'b' or file_path == 'B':
+                break
+
 
 def get_option():
     text = '> '
@@ -94,5 +155,13 @@ def get_option():
 
 
 def db_reset():
+    if not acc.active_account:
+        error_msg("You must log in first")
+        return
+    if not acc.active_account.admin:
+        error_msg('Only admin can clear Database ')
+        return
+
     db = mongoengine.connect(alias='core', name='mrsp')
     db.drop_database('mrsp')
+
